Tweak : Conversion center - aligned renamed link in bio core-render.php
